fix: don't track errors thrown from imported components
